+++ test/support/authentication.ex
@@ -80,6 +80,16 @@ defmodule TdBgWeb.Authentication do
     token
   end
 
+  def create_acl_entry(user_id, resource_type, resource_id, permissions) when is_list(permissions) do
+    MockPermissionResolver.create_acl_entry(%{
+      principal_type: "user",
+      principal_id: user_id,
+      resource_type: resource_type,
+      resource_id: resource_id,
+      permissions: permissions
+    })
+  end
+
   def create_acl_entry(user_id, resource_type, resource_id, role_name)
       when is_binary(role_name) do
     MockPermissionResolver.create_acl_entry(%{